requirements.txt
torch>=1.9.0
torchvision>=0.10.0
pandas>=1.3.0
numpy>=1.21.0
lifelines>=0.26.0
tqdm>=4.62.0
Pillow>=8.3.0
scikit-learn>=1.0.0
    